+++ lib/js_routes/generators/middleware.rb
@@ -1,15 +1,17 @@
-require "rails/generators"
+require "js_routes/generators/base"
 
-class JsRoutes::Generators::Middleware < Rails::Generators::Base
+class JsRoutes::Generators::Middleware < JsRoutes::Generators::Base
 
   source_root File.expand_path(__FILE__ + "/../../../templates")
 
   def create_middleware
     copy_file "initializer.rb", "config/initializers/js_routes.rb"
-    inject_into_file "app/javascript/packs/application.js", pack_content
     inject_into_file "config/environments/development.rb", middleware_content, before: /^end\n\z/
     inject_into_file "Rakefile", rakefile_content
     inject_into_file ".gitignore", gitignore_content
+    if path = application_js_path
+      inject_into_file path, pack_content
+    end
     JsRoutes.generate!
     JsRoutes.definitions!
   end